Add files via upload
The output of this code would be:

cat1: 1-20 = [19, 12, 7, 3, 2, 15, 20, 16, 11, 5]
Total numbers: 10
cat2: 21-40 = [35, 29, 39, 38, 25, 37, 33, 27, 24]
Total numbers: 9
cat3: 41-60 = [48, 53, 47, 51, 52, 50, 58, 60, 46, 49]
Total numbers: 10
cat4: 61-80 = [65, 77, 76, 75, 73, 74, 71, 67, 69, 80]
Total numbers: 10
cat5: 81-100 = [85, 89, 86, 97, 88, 92, 100, 91, 99, 95]
Total numbers: 10

In this example, the user entered an amount of 100, so the program generated 100 random numbers and classified them into the different categories (1-20, 21-40, 41-60, 61-80, 81-100). The program then printed out the numbers in each category and the total number of numbers in each category.